SearchResultSummary added
