+++ src/components/ResultPage/ResultPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CustomContainer } from '../CustomComponents/CustomContainer/CustomContainer';
-import { SearchResultSummary } from './SearchResultSummary';
+import { SearchResultSummary } from './SearchResultSummary/SearchResultSummary';
 
 const ResultPage = () => {
   return (
